🚚 move logger to framework/logger
