+++ test/e2e/lib/framework/createTest.ts
@@ -1,9 +1,9 @@
 import { deleteAllCookies, withBrowserLogs } from '../browserHelpers'
-import { log } from '../logger'
 import { flushEvents } from '../sdkHelpers'
 import { getTestServers, Servers, waitForServersIdle } from '../servers'
 import { EventRegistry } from './eventsRegistry'
 import { createIntakeServerApp } from './intakeServerApp'
+import { log } from './logger'
 import { createMockServerApp } from './mockServerApp'
 import { DEFAULT_SETUPS, LogsSetupOptions, RumSetupOptions, SetupFactory, SetupOptions } from './pageSetups'
 import { Endpoints } from './sdkBuilds'
