+++ test/e2e/lib/framework/logger.ts
